Refactor modmaker
Refactors modmaker to prepare for UI additions. Moves makemod into
a package. Reorganizes file modules. Moves all mod features to
packages. Moves randomizer to a package.
