--- internal/charStatsTxt/charStatsTxt.go
@@ -1,90 +1,0 @@
-package charStatsTxt
-
-// File Constants
-const (
-	FileName   = "CharStats.txt"
-	NumColumns = 41
-)
-
-// Header Indexes
-const (
-	Class              = 0
-	Str                = 1
-	Dex                = 2
-	Int                = 3
-	Vit                = 4
-	Tot                = 5
-	Stamina            = 6
-	HpAdd              = 7
-	PercentStr         = 8
-	PercentDex         = 9
-	PercentInt         = 10
-	PercentVit         = 11
-	ManaRegen          = 12
-	ToHitFactor        = 13
-	WalkVeLocity       = 14
-	RunVeLocity        = 15
-	RunDrain           = 16
-	Comment            = 17
-	LifePerLevel       = 18
-	StaminaPerLevel    = 19
-	ManaPerLevel       = 20
-	LifePerVitality    = 21
-	StaminaPerVitality = 22
-	ManaPerMagic       = 23
-	StatPerLevel       = 24
-	Walk               = 25
-	Run                = 26
-	Swing              = 27
-	Spell              = 28
-	GetHit             = 29
-	Bow                = 30
-	BLockFactor        = 31
-	StartSkill         = 32
-	Skill1             = 33
-	Skill2             = 34
-	Skill3             = 35
-	Skill4             = 36
-	Skill5             = 37
-	Skill6             = 38
-	Skill7             = 39
-	Skill8             = 40
-	Skill9             = 41
-	Skill10            = 42
-	StrAllSkills       = 43
-	StrSkillTab1       = 44
-	StrSkillTab2       = 45
-	StrSkillTab3       = 46
-	StrClassOnly       = 47
-	BaseWClass         = 48
-	Item1              = 49
-	Item1Loc           = 50
-	Item1Count         = 51
-	Item2              = 52
-	Item2Loc           = 53
-	Item2Count         = 54
-	Item3              = 55
-	Item3Loc           = 56
-	Item3Count         = 57
-	Item4              = 58
-	Item4Loc           = 59
-	Item4Count         = 60
-	Item5              = 61
-	Item5Loc           = 62
-	Item5Count         = 63
-	Item6              = 64
-	Item6Loc           = 65
-	Item6Count         = 66
-	Item7              = 67
-	Item7Loc           = 68
-	Item7Count         = 69
-	Item8              = 70
-	Item8Loc           = 71
-	Item8Count         = 72
-	Item9              = 73
-	Item9Loc           = 74
-	Item9Count         = 75
-	Item10             = 76
-	Item10Loc          = 77
-	Item10Count        = 78
-)
