--- internal/cubeMainTxt/cubeMainTxt.go
@@ -1,122 +1,0 @@
-package cubeMainTxt
-
-// File Constants
-const (
-	FileName   = "CubeMain.txt"
-	NumColumns = 105
-)
-
-// Recipe Descriptions
-const (
-	CowPortalWirt   = "1 Wirt's leg + 1 town portal book -> portal to cow lvl"
-	CowPortalNoWirt = "1 town portal scroll -> portal to cow lvl"
-)
-
-// Header Indexes
-const (
-	Description = 0
-	Enabled     = 1
-	Ladder      = 2
-	MinDiff     = 3
-	Version     = 4
-	Op          = 5
-	Param       = 6
-	Value       = 7
-	Class       = 8
-	NumInputs   = 9
-	Input1      = 10
-	Input2      = 11
-	Input3      = 12
-	Input4      = 13
-	Input5      = 14
-	Input6      = 15
-	Input7      = 16
-	Output      = 17
-	Lvl         = 18
-	Plvl        = 19
-	Ilvl        = 20
-	Mod1        = 21
-	Mod1Chance  = 22
-	Mod1Param   = 23
-	Mod1Min     = 24
-	Mod1Max     = 25
-	Mod2        = 26
-	Mod2Chance  = 27
-	Mod2Param   = 28
-	Mod2Min     = 29
-	Mod2Max     = 30
-	Mod3        = 31
-	Mod3Chance  = 32
-	Mod3Param   = 33
-	Mod3Min     = 34
-	Mod3Max     = 35
-	Mod4        = 36
-	Mod4Chance  = 37
-	Mod4Param   = 38
-	Mod4Min     = 39
-	Mod4Max     = 40
-	Mod5        = 41
-	Mod5Chance  = 42
-	Mod5Param   = 43
-	Mod5Min     = 44
-	Mod5Max     = 45
-	OutputB     = 46
-	BLvl        = 47
-	BPLvl       = 48
-	BILvl       = 49
-	BMod1       = 50
-	BMod1Chance = 51
-	BMod1Param  = 52
-	BMod1Min    = 53
-	BMod1Max    = 54
-	BMod2       = 55
-	BMod2Chance = 56
-	BMod2Param  = 57
-	BMod2Min    = 58
-	BMod2Max    = 59
-	BMod3       = 60
-	BMod3Chance = 61
-	BMod3Param  = 62
-	BMod3Min    = 63
-	BMod3Max    = 64
-	BMod4       = 65
-	BMod4Chance = 66
-	BMod4Param  = 67
-	BMod4Min    = 68
-	BMod4Max    = 69
-	BMod5       = 70
-	BMod5Chance = 71
-	BMod5Param  = 72
-	BMod5Min    = 73
-	BMod5Max    = 74
-	OutputC     = 75
-	CLvl        = 76
-	CPLvl       = 77
-	CILvl       = 78
-	CMod1       = 79
-	CMod1Chance = 80
-	CMod1Param  = 81
-	CMod1Min    = 82
-	CMod1Max    = 83
-	CMod2       = 84
-	CMod2Chance = 85
-	CMod2Param  = 86
-	CMod2Min    = 87
-	CMod2Max    = 88
-	CMod3       = 89
-	CMod3Chance = 90
-	CMod3Param  = 91
-	CMod3Min    = 92
-	CMod3Max    = 93
-	CMod4       = 94
-	CMod4Chance = 95
-	CMod4Param  = 96
-	CMod4Min    = 97
-	CMod4Max    = 98
-	CMod5       = 99
-	CMod5Chance = 100
-	CMod5Param  = 101
-	CMod5Min    = 102
-	CMod5Max    = 103
-	Eol         = 104
-)
