+++ internal/d2file/assets/asset_vars.go
