+++ internal/d2file/assets/assets_vfsdata.go
