+++ internal/d2file/txts/gems/gems.go
@@ -1,0 +1,52 @@
+package gems
+
+// File Constants
+const (
+	FileName   = "Gems.txt"
+	NumColumns = 41
+)
+
+// Header Indexes
+const (
+	Name            = 0
+	Letter          = 1
+	Transform       = 2
+	Code            = 3
+	NumMods         = 4
+	WeaponMod1Code  = 5
+	WeaponMod1Param = 6
+	WeaponMod1Min   = 7
+	WeaponMod1Max   = 8
+	WeaponMod2Code  = 9
+	WeaponMod2Param = 10
+	WeaponMod2Min   = 11
+	WeaponMod2Max   = 12
+	WeaponMod3Code  = 13
+	WeaponMod3Param = 14
+	WeaponMod3Min   = 15
+	WeaponMod3Max   = 16
+	HelmMod1Code    = 17
+	HelmMod1Param   = 18
+	HelmMod1Min     = 19
+	HelmMod1Max     = 20
+	HelmMod2Code    = 21
+	HelmMod2Param   = 22
+	HelmMod2Min     = 23
+	HelmMod2Max     = 24
+	HelmMod3Code    = 25
+	HelmMod3Param   = 26
+	HelmMod3Min     = 27
+	HelmMod3Max     = 28
+	ShieldMod1Code  = 29
+	ShieldMod1Param = 30
+	ShieldMod1Min   = 31
+	ShieldMod1Max   = 32
+	ShieldMod2Code  = 33
+	ShieldMod2Param = 34
+	ShieldMod2Min   = 35
+	ShieldMod2Max   = 36
+	ShieldMod3Code  = 37
+	ShieldMod3Param = 38
+	ShieldMod3Min   = 39
+	ShieldMod3Max   = 40
+)
