+++ internal/d2file/txts/levels/levels.go
@@ -1,0 +1,196 @@
+package levels
+
+// File Constants
+const (
+	FileName   = "Levels.txt"
+	NumColumns = 185
+)
+
+// Header Indexes
+const (
+	Name         = 0
+	Id           = 1
+	Pal          = 2
+	Act          = 3
+	QuestFlag    = 4
+	QuestFlagEx  = 5
+	Layer        = 6
+	SizeX        = 7
+	SizeY        = 8
+	SizeXN       = 9
+	SizeYN       = 10
+	SizeXH       = 11
+	SizeYH       = 12
+	OffsetX      = 13
+	OffsetY      = 14
+	Depend       = 15
+	Teleport     = 16
+	Rain         = 17
+	Mud          = 18
+	NoPer        = 19
+	LOSDraw      = 20
+	FloorFilter  = 21
+	BlankScreen  = 22
+	DrawEdges    = 23
+	IsInside     = 24
+	DrlgType     = 25
+	LevelType    = 26
+	SubType      = 27
+	SubTheme     = 28
+	SubWaypoint  = 29
+	SubShrine    = 30
+	Vis0         = 31
+	Vis1         = 32
+	Vis2         = 33
+	Vis3         = 34
+	Vis4         = 35
+	Vis5         = 36
+	Vis6         = 37
+	Vis7         = 38
+	Warp0        = 39
+	Warp1        = 40
+	Warp2        = 41
+	Warp3        = 42
+	Warp4        = 43
+	Warp5        = 44
+	Warp6        = 45
+	Warp7        = 46
+	Intensity    = 47
+	Red          = 48
+	Green        = 49
+	Blue         = 50
+	Portal       = 51
+	Position     = 52
+	SaveMonsters = 53
+	Quest        = 54
+	WarpDist     = 55
+	MonLvl1      = 56
+	MonLvl2      = 57
+	MonLvl3      = 58
+	MonLvl1Ex    = 59
+	MonLvl2Ex    = 60
+	MonLvl3Ex    = 61
+	MonDen       = 62
+	MonDenN      = 63
+	MonDenH      = 64
+	MonUMin      = 65
+	MonUMax      = 66
+	MonUMinN     = 67
+	MonUMaxN     = 68
+	MonUMinH     = 69
+	MonUMaxH     = 70
+	MonWndr      = 71
+	MonSpcWalk   = 72
+	NumMon       = 73
+	Mon1         = 74
+	Mon2         = 75
+	Mon3         = 76
+	Mon4         = 77
+	Mon5         = 78
+	Mon6         = 79
+	Mon7         = 80
+	Mon8         = 81
+	Mon9         = 82
+	Mon10        = 83
+	Mon11        = 84
+	Mon12        = 85
+	Mon13        = 86
+	Mon14        = 87
+	Mon15        = 88
+	Mon16        = 89
+	Mon17        = 90
+	Mon18        = 91
+	Mon19        = 92
+	Mon20        = 93
+	Mon21        = 94
+	Mon22        = 95
+	Mon23        = 96
+	Mon24        = 97
+	Mon25        = 98
+	RangedSpawn  = 99
+	NMon1        = 100
+	NMon2        = 101
+	NMon3        = 102
+	NMon4        = 103
+	NMon5        = 104
+	NMon6        = 105
+	NMon7        = 106
+	NMon8        = 107
+	NMon9        = 108
+	NMon10       = 109
+	NMon11       = 110
+	NMon12       = 111
+	NMon13       = 112
+	NMon14       = 113
+	NMon15       = 114
+	NMon16       = 115
+	NMon17       = 116
+	NMon18       = 117
+	NMon19       = 118
+	NMon20       = 119
+	NMon21       = 120
+	NMon22       = 121
+	NMon23       = 122
+	NMon24       = 123
+	NMon25       = 124
+	UMon1        = 125
+	UMon2        = 126
+	UMon3        = 127
+	UMon4        = 128
+	UMon5        = 129
+	UMon6        = 130
+	UMon7        = 131
+	UMon8        = 132
+	UMon9        = 133
+	UMon10       = 134
+	UMon11       = 135
+	UMon12       = 136
+	UMon13       = 137
+	UMon14       = 138
+	UMon15       = 139
+	UMon16       = 140
+	UMon17       = 141
+	UMon18       = 142
+	UMon19       = 143
+	UMon20       = 144
+	UMon21       = 145
+	UMon22       = 146
+	UMon23       = 147
+	UMon24       = 148
+	UMon25       = 149
+	CMon1        = 150
+	CMon2        = 151
+	CMon3        = 152
+	CMon4        = 153
+	CpCt1        = 154
+	CpCt2        = 155
+	CpCt3        = 156
+	CpCt4        = 157
+	CamT1        = 158
+	CamT2        = 159
+	CamT3        = 160
+	CamT4        = 161
+	Themes       = 162
+	SoundEnv     = 163
+	Waypoint     = 164
+	LevelName    = 165
+	LevelWarp    = 166
+	EntryFile    = 167
+	ObjGrp0      = 168
+	ObjGrp1      = 169
+	ObjGrp2      = 170
+	ObjGrp3      = 171
+	ObjGrp4      = 172
+	ObjGrp5      = 173
+	ObjGrp6      = 174
+	ObjGrp7      = 175
+	ObjPrb0      = 176
+	ObjPrb1      = 177
+	ObjPrb2      = 178
+	ObjPrb3      = 179
+	ObjPrb4      = 180
+	ObjPrb5      = 181
+	ObjPrb6      = 182
+	ObjPrb7      = 183
+	Beta         = 184
+)
