+++ internal/d2file/txts/magicSuffix/magicSuffix.go
@@ -1,0 +1,53 @@
+package magicSuffix
+
+// File Constants
+const (
+	FileName   = "MagicSuffix.txt"
+	NumColumns = 42
+)
+
+// Header Indexes
+const (
+	Name           = 0
+	Comment        = 1
+	Version        = 2
+	Spawnable      = 3
+	Rare           = 4
+	Level          = 5
+	MaxLevel       = 6
+	LevelReq       = 7
+	ClassSpecific  = 8
+	Class          = 9
+	ClassLevelReq  = 10
+	Frequency      = 11
+	Group          = 12
+	Mod1Code       = 13
+	Mod1Param      = 14
+	Mod1Min        = 15
+	Mod1Max        = 16
+	Mod2Code       = 17
+	Mod2Param      = 18
+	Mod2Min        = 19
+	Mod2Max        = 20
+	Mod3Code       = 21
+	Mod3Param      = 22
+	Mod3Min        = 23
+	Mod3Max        = 24
+	Transform      = 25
+	TransformColor = 26
+	IType1         = 27
+	IType2         = 28
+	IType3         = 29
+	IType4         = 30
+	IType5         = 31
+	IType6         = 32
+	IType7         = 33
+	EType1         = 34
+	EType2         = 35
+	EType3         = 36
+	EType4         = 37
+	EType5         = 38
+	Divide         = 39
+	Multiply       = 40
+	Add            = 41
+)
