+++ internal/d2file/txts/misc/misc.go
@@ -1,0 +1,189 @@
+package misc
+
+// File Constants
+const (
+	FileName   = "Misc.txt"
+	NumColumns = 169
+)
+
+// Misc Names
+const (
+	IdentifyBook   = "Identify Book"
+	TownPortalBook = "Town Portal Book"
+	Arrows         = "Arrows"
+	Bolts          = "Bolts"
+	SkeletonKey    = "Skeleton Key"
+)
+
+// Header Indexes
+const (
+	Name             = 0
+	Name_            = 1
+	SZFlavorText     = 2
+	CompactSave      = 3
+	Version          = 4
+	Level            = 5
+	LevelReq         = 6
+	Rarity           = 7
+	Spawnable        = 8
+	Speed            = 9
+	NoDurability     = 10
+	Cost             = 11
+	GambleCost       = 12
+	AutoPrefix       = 13
+	Code             = 14
+	AlternateGfx     = 15
+	NameStr          = 16
+	Component        = 17
+	InvWidth         = 18
+	InvHeight        = 19
+	HasInv           = 20
+	GemSockets       = 21
+	GemApplyType     = 22
+	FlippyFile       = 23
+	InvFile          = 24
+	UniqueInvFile    = 25
+	Special          = 26
+	Transmogrify     = 27
+	TMogType         = 28
+	TMogMin          = 29
+	TMogMax          = 30
+	Useable          = 31
+	Throwable        = 32
+	Type_            = 33
+	Type2            = 34
+	DropSound        = 35
+	DropsFxFrame     = 36
+	UseSound         = 37
+	Unique           = 38
+	Transparent      = 39
+	TransTbl         = 40
+	LightRadius      = 41
+	Belt             = 42
+	AutoBelt         = 43
+	Stackable        = 44
+	MinStack         = 45
+	MaxStack         = 46
+	SpawnStack       = 47
+	Quest            = 48
+	QuestDiffCheck   = 49
+	MissileType      = 50
+	SpellIcon        = 51
+	PSpell           = 52
+	State            = 53
+	Cstate1          = 54
+	Cstate2          = 55
+	Len              = 56
+	Stat1            = 57
+	Calc1            = 58
+	Stat2            = 59
+	Calc2            = 60
+	Stat3            = 61
+	Calc3            = 62
+	SpellDesc        = 63
+	SpellDescStr     = 64
+	SpellDescCalc    = 65
+	DurWarning       = 66
+	QntWarning       = 67
+	GemOffset        = 68
+	BetterGem        = 69
+	BitField1        = 70
+	CharsiMin        = 71
+	CharsiMax        = 72
+	CharsiMagicMin   = 73
+	CharsiMagicMax   = 74
+	CharsiMagicLvl   = 75
+	GheedMin         = 76
+	GheedMax         = 77
+	GheedMagicMin    = 78
+	GheedMagicMax    = 79
+	GheedMagicLvl    = 80
+	AkaraMin         = 81
+	AkaraMax         = 82
+	AkaraMagicMin    = 83
+	AkaraMagicMax    = 84
+	AkaraMagicLvl    = 85
+	FaraMin          = 86
+	FaraMax          = 87
+	FaraMagicMin     = 88
+	FaraMagicMax     = 89
+	FaraMagicLvl     = 90
+	LysanderMin      = 91
+	LysanderMax      = 92
+	LysanderMagicMin = 93
+	LysanderMagicMax = 94
+	LysanderMagicLvl = 95
+	DrognanMin       = 96
+	DrognanMax       = 97
+	DrognanMagicMin  = 98
+	DrognanMagicMax  = 99
+	DrognanMagicLvl  = 100
+	HraltiMin        = 101
+	HraltiMax        = 102
+	HraltiMagicMin   = 103
+	HraltiMagicMax   = 104
+	HraltiMagicLvl   = 105
+	AlkorMin         = 106
+	AlkorMax         = 107
+	AlkorMagicMin    = 108
+	AlkorMagicMax    = 109
+	AlkorMagicLvl    = 110
+	OrmusMin         = 111
+	OrmusMax         = 112
+	OrmusMagicMin    = 113
+	OrmusMagicMax    = 114
+	OrmusMagicLvl    = 115
+	ElzixMin         = 116
+	ElzixMax         = 117
+	ElzixMagicMin    = 118
+	ElzixMagicMax    = 119
+	ElzixMagicLvl    = 120
+	AshearaMin       = 121
+	AshearaMax       = 122
+	AshearaMagicMin  = 123
+	AshearaMagicMax  = 124
+	AshearaMagicLvl  = 125
+	CainMin          = 126
+	CainMax          = 127
+	CainMagicMin     = 128
+	CainMagicMax     = 129
+	CainMagicLvl     = 130
+	HalbuMin         = 131
+	HalbuMax         = 132
+	HalbuMagicMin    = 133
+	HalbuMagicMax    = 134
+	HalbuMagicLvl    = 135
+	MalahMin         = 136
+	MalahMax         = 137
+	MalahMagicMin    = 138
+	MalahMagicMax    = 139
+	MalahMagicLvl    = 140
+	LarzukMin        = 141
+	LarzukMax        = 142
+	LarzukMagicMin   = 143
+	LarzukMagicMax   = 144
+	LarzukMagicLvl   = 145
+	DrehyaMin        = 146
+	DrehyaMax        = 147
+	DrehyaMagicMin   = 148
+	DrehyaMagicMax   = 149
+	DrehyaMagicLvl   = 150
+	JamellaMin       = 151
+	JamellaMax       = 152
+	JamellaMagicMin  = 153
+	JamellaMagicMax  = 154
+	JamellaMagicLvl  = 155
+	SourceArt        = 156
+	GameArt          = 157
+	Transform        = 158
+	InvTrans         = 159
+	SkipName         = 160
+	NightmareUpgrade = 161
+	HellUpgrade      = 162
+	mindam           = 163
+	MaxDam           = 164
+	PermStoreItem    = 165
+	Multibuy         = 166
+	Nameable         = 167
+	Eol              = 168
+)
