+++ internal/d2file/txts/missiles/missiles.go
@@ -1,0 +1,182 @@
+package missiles
+
+// File Constants
+const (
+	FileName   = "Missiles.txt"
+	NumColumns = 171
+)
+
+// Header Indexes
+const (
+	Missile             = 0
+	Id                  = 1
+	PCltDoFunc          = 2
+	PCltHitFunc         = 3
+	PSrvDoFunc          = 4
+	PSrvHitFunc         = 5
+	PSrvDmgFunc         = 6
+	SrvCalc1            = 7
+	SrvCalc1Desc        = 8
+	Param1              = 9
+	Param1Desc          = 10
+	Param2              = 11
+	Param2Desc          = 12
+	Param3              = 13
+	Param3Desc          = 14
+	Param4              = 15
+	Param4Desc          = 16
+	Param5              = 17
+	Param5Desc          = 18
+	CltCalc1            = 19
+	ClientCalc1Desc     = 20
+	CltParam1           = 21
+	ClientParam1Desc    = 22
+	CltParam2           = 23
+	ClientParam2Desc    = 24
+	CltParam3           = 25
+	ClientParam3Desc    = 26
+	CltParam4           = 27
+	ClientParam4Desc    = 28
+	CltParam5           = 29
+	ClientParam5Desc    = 30
+	SHitCalc1           = 31
+	ServerHitCalc1Desc  = 32
+	SHitPar1            = 33
+	ServerHitParam1Desc = 34
+	SHitPar2            = 35
+	ServerHitParam2Desc = 36
+	SHitPar3            = 37
+	ServerHitParam3Desc = 38
+	CHitCalc1           = 39
+	ClientHitCalc1Desc  = 40
+	CHitPar1            = 41
+	ClientHitParam1Desc = 42
+	CHitPar2            = 43
+	ClientHitParam2Desc = 44
+	CHitPar3            = 45
+	ClientHitParam3Desc = 46
+	DmgCalc1            = 47
+	DamageCalc1         = 48
+	DParam1             = 49
+	DamageParam1Desc    = 50
+	DParam2             = 51
+	DamageParam2Desc    = 52
+	Vel                 = 53
+	MaxVel              = 54
+	VelLev              = 55
+	Accel               = 56
+	Range               = 57
+	LevRange            = 58
+	Light               = 59
+	Flicker             = 60
+	Red                 = 61
+	Green               = 62
+	Blue                = 63
+	InitSteps           = 64
+	Activate            = 65
+	LoopAnim            = 66
+	CelFile             = 67
+	animrate            = 68
+	AnimLen             = 69
+	AnimSpeed           = 70
+	RandStart           = 71
+	SubLoop             = 72
+	SubStart            = 73
+	SubStop             = 74
+	CollideType         = 75
+	CollideKill         = 76
+	CollideFriend       = 77
+	LastCollide         = 78
+	Collision           = 79
+	ClientCol           = 80
+	ClientSend          = 81
+	NextHit             = 82
+	NextDelay           = 83
+	XOffset             = 84
+	YOffset             = 85
+	ZOffset             = 86
+	Size                = 87
+	SrcTown             = 88
+	CltSrcTown          = 89
+	CanDestroy          = 90
+	ToHit               = 91
+	AlwaysExplode       = 92
+	Explosion           = 93
+	Town                = 94
+	NoUniqueMod         = 95
+	NoMultiShot         = 96
+	Holy                = 97
+	CanSlow             = 98
+	ReturnFire          = 99
+	GetHit              = 100
+	SoftHit             = 101
+	KnockBack           = 102
+	Trans               = 103
+	Qty                 = 104
+	Pierce              = 105
+	SpecialSetup        = 106
+	MissileSkill        = 107
+	Skill               = 108
+	ResultFlags         = 109
+	HitFlags            = 110
+	HitShift            = 111
+	ApplyMastery        = 112
+	SrcDamage           = 113
+	Half2HSrc           = 114
+	SrcMissDmg          = 115
+	MinDamage           = 116
+	MinLevDam1          = 117
+	MinLevDam2          = 118
+	MinLevDam3          = 119
+	MinLevDam4          = 120
+	MinLevDam5          = 121
+	MaxDamage           = 122
+	MaxLevDam1          = 123
+	MaxLevDam2          = 124
+	MaxLevDam3          = 125
+	MaxLevDam4          = 126
+	MaxLevDam5          = 127
+	DmgSymPerCalc       = 128
+	EType               = 129
+	EMin                = 130
+	MinELev1            = 131
+	MinELev2            = 132
+	MinELev3            = 133
+	MinELev4            = 134
+	MinELev5            = 135
+	Emax                = 136
+	MaxELev1            = 137
+	MaxELev2            = 138
+	MaxELev3            = 139
+	MaxELev4            = 140
+	MaxELev5            = 141
+	EDmgSymPerCalc      = 142
+	ELen                = 143
+	ELevLen1            = 144
+	ELevLen2            = 145
+	ELevLen3            = 146
+	HitClass            = 147
+	NumDirections       = 148
+	LocalBlood          = 149
+	DamageRate          = 150
+	TravelSound         = 151
+	HitSound            = 152
+	ProgSound           = 153
+	ProgOverlay         = 154
+	ExplosionMissile    = 155
+	SubMissile1         = 156
+	SubMissile2         = 157
+	SubMissile3         = 158
+	HitSubMissile1      = 159
+	HitSubMissile2      = 160
+	HitSubMissile3      = 161
+	HitSubMissile4      = 162
+	CltSubMissile1      = 163
+	CltSubMissile2      = 164
+	CltSubMissile3      = 165
+	CltHitSubMissile1   = 166
+	CltHitSubMissile2   = 167
+	CltHitSubMissile3   = 168
+	CltHitSubMissile4   = 169
+	EOL                 = 170
+)
