+++ internal/d2file/txts/properties/properties.go
@@ -1,0 +1,47 @@
+package properties
+
+// File Constants
+const (
+	FileName   = "Properties.txt"
+	NumColumns = 36
+)
+
+// Header Indexes
+const (
+	Code  = 0
+	Done  = 1
+	Set1  = 2
+	Val1  = 3
+	Func1 = 4
+	Stat1 = 5
+	Set2  = 6
+	Val2  = 7
+	Func2 = 8
+	Stat2 = 9
+	Set3  = 10
+	Val3  = 11
+	Func3 = 12
+	Stat3 = 13
+	Set4  = 14
+	Val4  = 15
+	Func4 = 16
+	Stat4 = 17
+	Set5  = 18
+	Val5  = 19
+	Func5 = 20
+	Stat5 = 21
+	Set6  = 22
+	Val6  = 23
+	Func6 = 24
+	Stat6 = 25
+	Set7  = 26
+	Val7  = 27
+	Func7 = 28
+	Stat7 = 29
+	Desc  = 30
+	Param = 31
+	Min   = 32
+	Max   = 33
+	Notes = 34
+	Eol   = 35
+)
