+++ internal/d2file/txts/treasureClassEx/treasureClassEx.go
@@ -1,0 +1,54 @@
+package treasureClassEx
+
+// File Constants
+const (
+	FileName   = "TreasureClassEx.txt"
+	NumColumns = 33
+)
+
+// Row Constants
+const (
+	Andariel   = "Andariel"
+	Duriel     = "Duriel"
+	DurielBase = "Duriel - Base"
+	Mephisto   = "Mephisto"
+	Diablo     = "Diablo"
+	Baal       = "Baal"
+)
+
+// Header Indexes
+const (
+	TreasureClass = 0
+	Group         = 1
+	Level         = 2
+	Picks         = 3
+	Unique        = 4
+	Set           = 5
+	Rare          = 6
+	Magic         = 7
+	NoDrop        = 8
+	Item1         = 9
+	Prob1         = 10
+	Item2         = 11
+	Prob2         = 12
+	Item3         = 13
+	Prob3         = 14
+	Item4         = 15
+	Prob4         = 16
+	Item5         = 17
+	Prob5         = 18
+	Item6         = 19
+	Prob6         = 20
+	Item7         = 21
+	Prob7         = 22
+	Item8         = 23
+	Prob8         = 24
+	Item9         = 25
+	Prob9         = 26
+	Item10        = 27
+	Prob10        = 28
+	SumItems      = 29
+	TotalProb     = 30
+	DropChance    = 31
+	Term          = 32
+)
