--- internal/itemRatioTxt/itemRatioTxt.go
@@ -1,31 +1,0 @@
-package itemRatioTxt
-
-// File Constants
-const (
-	FileName   = "ItemRatio.txt"
-	NumColumns = 20
-)
-
-// Header Indexes
-const (
-	Function         = 0
-	Version          = 1
-	Uber             = 2
-	ClassSpecific    = 3
-	Unique           = 4
-	UniqueDivisor    = 5
-	UniqueMin        = 6
-	Rare             = 7
-	RareDivisor      = 8
-	RareMin          = 9
-	Set              = 10
-	SetDivisor       = 11
-	SetMin           = 12
-	Magic            = 13
-	MagicDivisor     = 14
-	MagicMin         = 15
-	HiQuality        = 16
-	HiQualityDivisor = 17
-	Normal           = 18
-	NormalDivisor    = 19
-)
