--- internal/itemStatCostTxt/itemStatCostTxt.go
@@ -1,64 +1,0 @@
-package itemStatCostTxt
-
-// File Constants
-const (
-	FileName   = "ItemStatCost.txt"
-	NumColumns = 53
-)
-
-// Header Indexes
-const (
-	Stat           = 0
-	ID             = 1
-	SendOther      = 2
-	Signed         = 3
-	SendBits       = 4
-	SendParamBits  = 5
-	UpdateAnimRate = 6
-	Saved          = 7
-	CSvSigned      = 8
-	CSvBits        = 9
-	CSvParam       = 10
-	fCallback      = 11
-	fMin           = 12
-	MinAccr        = 13
-	Encode         = 14
-	Add            = 15
-	Multiply       = 16
-	Divide         = 17
-	ValShift       = 18
-	SaveBits109    = 19
-	SaveAdd109     = 20
-	SaveBits       = 21
-	SaveAdd        = 22
-	SaveParamBits  = 23
-	keepzero       = 24
-	Op             = 25
-	OpParam        = 26
-	OpBase         = 27
-	OpStat1        = 28
-	OpStat2        = 29
-	OpStat3        = 30
-	Direct         = 31
-	MaxStat        = 32
-	ItemSpecific   = 33
-	DamageRelated  = 34
-	ItemEvent1     = 35
-	ItemEventfunc1 = 36
-	ItemEvent2     = 37
-	ItemEventfunc2 = 38
-	Descpriority   = 39
-	Descfunc       = 40
-	Descval        = 41
-	Descstrpos     = 42
-	Descstrneg     = 43
-	Descstr2       = 44
-	Dgrp           = 45
-	Dgrpfunc       = 46
-	Dgrpval        = 47
-	Dgrpstrpos     = 48
-	Dgrpstrneg     = 49
-	Dgrpstr2       = 50
-	Dtuff          = 51
-	Eol            = 52
-)
