--- internal/runesTxt/runesTxt.go
@@ -1,61 +1,0 @@
-package runesTxt
-
-// File Constants
-const (
-	FileName    = "Runes.txt"
-	NumColumns  = 49
-	MaxNumProps = 7
-)
-
-// Header Indexes
-const (
-	Name     = 0
-	RuneName = 1
-	Complete = 2
-	Server   = 3
-	IType1   = 4
-	IType2   = 5
-	IType3   = 6
-	IType4   = 7
-	IType5   = 8
-	IType6   = 9
-	EType1   = 10
-	EType2   = 11
-	EType3   = 12
-	Runes    = 13
-	Rune1    = 14
-	Rune2    = 15
-	Rune3    = 16
-	Rune4    = 17
-	Rune5    = 18
-	Rune6    = 19
-	T1Code1  = 20
-	T1Param1 = 21
-	T1Min1   = 22
-	T1Max1   = 23
-	T1Code2  = 24
-	T1Param2 = 25
-	T1Min2   = 26
-	T1Max2   = 27
-	T1Code3  = 28
-	T1Param3 = 29
-	T1Min3   = 30
-	T1Max3   = 31
-	T1Code4  = 32
-	T1Param4 = 33
-	T1Min4   = 34
-	T1Max4   = 35
-	T1Code5  = 36
-	T1Param5 = 37
-	T1Min5   = 38
-	T1Max5   = 39
-	T1Code6  = 40
-	T1Param6 = 41
-	T1Min6   = 42
-	T1Max6   = 43
-	T1Code7  = 44
-	T1Param7 = 45
-	T1Min7   = 46
-	T1Max7   = 47
-	Eol      = 48
-)
