--- internal/superUniquesTxt/superUniquesTxt.go
@@ -1,37 +1,0 @@
-package superUniquesTxt
-
-// File Constants
-const (
-	FileName   = "SuperUniques.txt"
-	NumColumns = 21
-)
-
-// Super Uniques
-const (
-	CowKing = "The Cow King"
-)
-
-// Header Indexes
-const (
-	Superunique = 0
-	Name        = 1
-	Class       = 2
-	HcIdx       = 3
-	MonSound    = 4
-	Mod1        = 5
-	Mod2        = 6
-	Mod3        = 7
-	MinGrp      = 8
-	MaxGrp      = 9
-	EClass      = 10
-	AutoPos     = 11
-	Stacks      = 12
-	Replaceable = 13
-	Utrans      = 14
-	UtransN     = 15
-	UtransH     = 16
-	TC          = 17
-	TCN         = 18
-	TCH         = 19
-	Eol         = 20
-)
